embeddings/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
numpy>=1.24.0
pymysql>=1.1.0
beautifulsoup4>=4.12.0
